internal server implementation for city
not in the networking sense of course
currently does not stutter, since life is soooo much simpler without stuttering now that things are multithreaded.
stuttering is an important feature though, and should be added in some form.
Not sure whether to finely or coarsely grain the rwlocks for stuttering, but it would be hard to bench until I have something substantial to test.
all in all, stuttering can wait.